[WFLY-20628] Match package names in tests where it makes sense, so they match the packages in the main.
